feat: Add layout for header and banner with conditional rendering
