+++ homework-intermediate/day30/src/app/boards/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import BoardList from '@/components/boards-list/list';
+import BoardList from '@/components/boards-list';
 
 export default function BoardsPage() {
   return <BoardList />;
