+++ homework-intermediate/day30/src/components/boards-list/styles.module.css
@@ -2,11 +2,11 @@
 
 .boardBody {
   display: flex;
-  padding: 0px 320px;
   flex-direction: column;
   align-items: center;
   gap: 40px;
-  align-self: stretch;
+  max-width: 1280px;
+  width: 100%;
 }
 
 .boardInnerBody {
@@ -16,14 +16,12 @@
 .boardFrame {
   margin-top: 40px;
   display: flex;
-  max-width: 1280px;
-  min-width: 640px;
+  width: 100%;
   flex-direction: column;
   align-items: center;
   gap: 24px;
   align-self: stretch;
   padding: 24px 48px;
-
   box-shadow: 0px 0px 20px rgba(0, 0, 0, 0.08);
   border-radius: 16px;
 }
