+++ homework-intermediate/day30/src/components/boards-list/types.ts
