+++ homework-intermediate/day30/tailwind.config.ts
@@ -5,6 +5,7 @@ const config: Config = {
     './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
     './src/components/**/*.{js,ts,jsx,tsx,mdx}',
     './src/app/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/commons/**/*.{js,ts,jsx,tsx,mdx}',
   ],
   theme: {
     extend: {
